refactor - cleanup
