+++ src/ts/waterfall/svg-row-subcomponents.ts
